extract the analyzer from the stream
