+++ test/semantic-types.test.ts
@@ -2,7 +2,7 @@ import assert from 'assert';
 import BSON from 'bson';
 
 import getSchema from '../src';
-import type { Schema, PrimitiveSchemaType } from '../src/stream';
+import type { Schema, PrimitiveSchemaType } from '../src/schema-analyzer';
 
 describe('options', function() {
   const docs = [
